Cobatl Strike/Cobalt Strike Dec/requirements.txt
pefile
peutils
minidump
pyzipper
colorama
pycryptodome
hexdump
